Separate subparser stuff into CLI module
