+++ kevlar/cli/reaugment.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+#
+# -----------------------------------------------------------------------------
+# Copyright (c) 2017 The Regents of the University of California
+#
+# This file is part of kevlar (http://github.com/dib-lab/kevlar) and is
+# licensed under the MIT license: see LICENSE.
+# -----------------------------------------------------------------------------
+
+
+def subparser(subparsers):
+    subparser = subparsers.add_parser('reaugment')
+    subparser.add_argument('-o', '--out', metavar='FILE',
+                           help='output file; default is terminal (stdout)')
+    subparser.add_argument('augfastq', help='original augmented Fastq file')
+    subparser.add_argument('fastq', help='processed Fastq file to re-annotate')
